chore: API폴더 명 및 함수 명 변경
