+++ apis/LikeAndCommentAPI.ts
@@ -8,11 +8,3 @@ export const addLikeProject = async (id: number) => {
   const response = await axios.post(`/like/${id}`, null, config);
   return response;
 };
-export const addLikeRecruit = async (id: number) => {
-  const token = localStorage.getItem("accessToken");
-  const config = {
-    headers: { Authorization: `Bearer ${token}` },
-  };
-  const response = await axios.post(`/recruit-board/likes/${id}`, null, config);
-  return response;
-};
