--- components/Alarm/AlarmQurey.tsx
@@ -1,1 +1,0 @@
-
