+++ hooks/mutations/useAddLikeProject.tsx
@@ -1,12 +1,12 @@
 import { useMutation, useQueryClient } from "@tanstack/react-query";
-import { addLikeRecruit } from "apis/LikeBoardApi";
+import { addLikeProject } from "apis/LikeAndCommentAPI";
 import useToast from "../common/useToast";
 
-export const useLikeRecruit = () => {
+export const useAddLikeProject = () => {
   const queryClient = useQueryClient();
   const { addToast } = useToast();
   const { mutate } = useMutation({
-    mutationFn: addLikeRecruit,
+    mutationFn: addLikeProject,
     onSuccess: (res) => {
       if (res.data.message.includes("추천했습니다")) {
         addToast({
@@ -15,7 +15,7 @@ export const useLikeRecruit = () => {
           content: "관심게시물로 등록되었습니다",
         });
       }
-      queryClient.invalidateQueries({ queryKey: ["recruits"] });
+      queryClient.invalidateQueries({ queryKey: ["projectData"] });
       queryClient.invalidateQueries({ queryKey: ["mypageData"] });
     },
   });
