+++ hooks/queries/useGetMypageData.tsx
@@ -1,5 +1,5 @@
 import { useQuery } from "@tanstack/react-query";
-import { getMypageData } from "apis/MypageApi";
+import { getMypageData } from "apis/UserAPI";
 import useToast from "../common/useToast";
 export const useGetMypageData = () => {
   const { addToast, deleteToast } = useToast();
